Feature/Leo wallet authentication (#21)
## Description

This PR introduces the login flow using the Leo wallet. We can verify
that the user is the owner of the private key by verifying the signature
produced by the Leo wallet. When the signature is verified we produce a
JWT token that is used in the frontend as authentication token.
 
## Related Issues

#20 

## Changes Made

- Add Aleo SDK 
- Add **verifySignature** function for verifying the signature produced
by the Leo wallet
- Add logic that requests the Leo wallet to sign a message
- Add **verifySignature** function to the nextAuth **authorise**
function