Feat: Update Chat Stage, Voting Stage of the Match (#182)
## Description
This pull request introduces UI updates with the primary focus on the
Match state of the game. Here's a breakdown of the key updates:

## Changes

### Chat Stage
During the chat stage, each user is assigned a character and may begin
chatting with each other. A prompt provided by the host is used to start
conversation, with the main goal of determining who is the bot. The chat
stage has a limited time.

- Character Avatars: Added Pixel art assigned for each character.
- Host Prompt: Updated the host prompt to the top of the message
component.
- Input Field UI styling and added multiline capability for writing long
messages

### Voting Stage
- Players can vote who they think is the bot by selecting multiple
characters

## Other Changes
- Improved Pixel Art Quality: Eliminated unwanted strokes in pixel
images introduced by Figma.
- Match Feature Layout: We've overhauled the match feature layout for
better usability.
- SVGR Plugin Reconfiguration: Reconfigured the SVGR plugin responsible
for rendering SVGs as React components, improving performance and
reliability.
- Main Layout UI Tweaks: Made subtle tweaks to the main layout UI for
improved aesthetics and usability.
- Custom Scrollbar
- Theme settings
- Primary Button
- Game Timer Feature ( Used in Chat and Voting Stage)