feat(support page): integration and validation (#219)
closes #218 

- [x] refactor validation logic to be shared in fe and be
- [x] track error state in support page
- [x] implement error styling in input components
- [x] test end-to-end

[Zod](https://zod.dev/) schemas are defined in
`src/constants/validation.ts` and used both in frontend and backend
code, along with relevant types.

At the component level input validation is tracked independently for
each field, triggered by `onBlur` and if an error is present also by
`onChange`. Send is disabled when an error is present. In the be, the
schema is passed to the `supportRouter` for validation.

No errors are thrown for the topic select at the moment, they may choose
to omit it or pick one that fits their inquiry.

![Screenshot 2024-01-25 at 15 04
51](https://github.com/Kryha/bot-busters/assets/97608823/1524a537-7ec9-4f16-964c-639b4e104347)